Introduced 'setattr' method adn updated test suite to use it.
NOTE: This is an intermediate version allowing seamless update of config files to the next version.

Next version will change the attribute setting syntax from:
    item.someattribute(env1=x,...,envN=y)
to:
    item.setattr('someattribute', env1=x,...,envN=y)

This change is necessary in order to avoid a lot of problems with frozen/not-frozen access to attributes, as it will allow automatic freeze on __getattr__ call.